add all repositories

+++ service/model/book.go
@@ -3,7 +3,7 @@ package model
 type Book struct {
 	Id            int64   `json:"id"`
 	YearPublished int     `json:"yearPublished"`
-	Rating        int     `json:"rating"`
+	Rating        float32 `json:"rating"`
 	Pages         int     `json:"pages"`
 	Title         string  `json:"title"`
 	Genre         *Genre  `json:"genre"`
